refactor(provider): simplify CCIP Read client and HTTP gateway
Split the reqwest gateway into its own module, read bodies with
Response::chunk so the provider's reqwest dependency no longer needs the
json and stream features, and accept http gateways alongside https so
local gateways work. Default CCIP calls to the latest block like plain
eth_call, drop the redundant Arc/unreachable plumbing, and expose the
gateway through the alloy meta crate's provider-ccip-read feature.
